Added new partner customization to load wallpapers
Lollipop introduced the ability to load 3rd party wallpapers provided
that those apks are system apps.  All that is required is to declare
a receiver in the manifest that handles the action
com.android.launcher3.action.PARTNER_CUSTOMIZATION.  This receiver
is never launched so there is no need to have a source file for it.

And added a set wallpaper button for nowallpaper
We set the wallpaper to black without a confirm
button currently and without exit the wallpaper picker.
This causes odd behaviour.
Add back a confirm button like live wallpaper picker